Python3_HomeWork_Chat/Lesson_8_ Streams/Lesson/requirements.txt
ipython
jupyter
chardet
pyyaml
pytest
pytest-cov
